% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_engine.R
\name{cachesuite_engine}
\alias{cachesuite_engine}
\title{Knitr language engine for state-aware caching evaluation of R code}
\usage{
cachesuite_engine(options)
}
\arguments{
\item{options}{Options object passed by knitr to language engine}
}
\description{
This function defines an alternative knitr language engine
for evaluating R code using RCacheSuite. It can be specified
as the engine for a particular 'language' /chunk type via knitr, but
should never be called directly by end users.
}
