\name{CachingEngine-class}
\Rdversion{1.1}
\docType{class}
\alias{CachingEngine-class}
\alias{show,CachingEngine-method}
\title{Class \code{"CachingEngine"}}
\description{
  A reference class representing a custom caching
  mechanism, including the location caches should be stored and
  functions to perform both code evaluation and the processing of values
  returned by evaluation.
}
\section{Fields}{
  \describe{
    \item{\code{base_dir}:}{\code{character}. Base directory to look for/create code-level
      cache sets in.}
    \item{\code{tmp_base_dir}:}{\code{character}. Temporary base directory to look
      for/create temporary caches in.}
    \item{\code{cache_sets}:}{\code{list}. A list containing the \code{CodeCacheSet}
      objects this caching engine currently knows about.}
    \item{\code{gdevs}:}{\code{list}. A list of functions to be used to create image
      files of any graphical output generated during evaluation of code.}

    \item{\code{write_allowed}:}{\code{logical}. Whether this caching
      engine supports writing caches to disk.}
    \item{\code{write_on_cache}:}{\code{logical}. Whether every cache should be written
      to disk as soon as it is created (i.e. directly after evaluation
      of code which is eligable for caching.}
    \item{\code{eval_fun}:}{\code{function}. The function which will be
      used to evaluate code when this caching engine is in use. Must
      accept \code{code}, \code{env}, and \code{\dots}. Its return value
      will be passed to \code{return_handler}}
    \item{\code{return_handler}:}{\code{function}. A handler for the
      objects returned by \code{eval_fun}. Must accept \code{val},
      \code{graphics}, code{evaled} (indicating whether \code{val}
      was generated by evaluation or loaded from a cache), and \code{\dots}}
  }
}
\section{Class-Based Methods}{
  \describe{
    \item{\code{to_disk(clear_mem, tmp)}:}{ Write the contents of all
      code-level caches to disk. If \code{clear_mem} is \code{TRUE},
      remove the objects from memory after a succesful write. If
      \code{tmp} is \code{TRUE}, write the cache sets to
      \code{tmp_base_dir} instead of \code{base_dir}}
  }
}
\section{Constructor}{
  The \code{\link{cachingEngine}} function is provided as a constructor for
  this class with reasonable default values for the fields.
}
\author{
  Gabriel Becker
}
\seealso{
  \code{\link{cachingEngine}} \code{\link{evalWithCache}} \code{\linkS4class{CodeCacheSet} }
}
\keyword{classes}
\keyword{programming}