% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\name{cachingEngine}
\alias{cachingEngine}
\title{CachingEngine Constructor}
\usage{
cachingEngine(base_dir = "./r_caches", tmp_base_dir = tempdir(),
  write_allowed = TRUE, write_on_cache = FALSE, gdevs = list(png),
  eval_fun = parseWithVis, return_handler = withVisHandler,
  populate = TRUE)
}
\arguments{
\item{base_dir}{character value . Path to which permanent caches
should be written (if applicable)}

\item{tmp_base_dir}{character value. Path to which temporary caches
should be written (if applicable)}

\item{write_allowed}{logical. Should the engine allow caches to be
written to disk. Defaults to \code{TRUE}}

\item{write_on_cache}{logical. Should the engine write caches to
disk immediately upon creation. Defaults to \code{FALSE}}

\item{gdevs}{list. A list of graphics device functions to use when
caching plots. Defaults to \code{list(png)}}

\item{eval_fun}{function. The evaluator function to be used when no
cache is found. Defaults to \code{parseWithVis}.}

\item{return_handler}{function. The function to process values
returned by \code{eval_fun}. This function will be called on
values returned from \code{eval_fun} directly and those loaded
from cache. Defaults to \code{withVisHandler} which
supports plot caching.}

\item{populate}{logical. Should the engine include existing caches
from \code{base_dir} upon creation. Defaults to \code{TRUE}}
}
\value{
A CachingEngine object suitable for use with
    \code{evalWithCache}
}
\description{
Function for creating a Caching engine with specific
    behaviors
}
\seealso{
\code{\link{evalWithCache}}
}
\author{
Gabriel Becker
}
