% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prov.R
\name{fullprovgraph}
\alias{fullprovgraph}
\title{Create the full (multiple values per variable) provenance
    graph for a cache}
\usage{
fullprovgraph(cacheobj, df = cacheobj$provstore)
}
\arguments{
\item{cacheobj}{a CachingEngine or CodeSetCache object. Ignored if
\code{df} is set.}

\item{df}{ProvStoreDF (or appropriately formatted
data.frame). Generally \code{cacheobj} should be specified and
the default of \code{cacheobj$provstore} should be used.}
}
\value{
an igraph object representing the provenance graph
}
\description{
This generates and returns the \emph{full} provenance
    graph reflecting the information stored in the cache or
    data.frame specified. This can include the same variable
    multiple times if the corresponding expression is run with
    different inputs.
}
