% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_support.R
\name{make_evaluate_ef}
\alias{make_evaluate_ef}
\title{Make an 'evaluate'-based evaluation function}
\usage{
make_evaluate_ef(source = NULL, text = NULL, graphics = NULL,
  message = NULL, warning = NULL, error = NULL, value = NULL)
}
\arguments{
\item{source}{An additional source handler to call before the one
we use internally, or NULL (default)}

\item{text}{An additional text handler to call before the one we
use internally, or NULL (default)}

\item{graphics}{An additional graphics handler to call before the
one we use internally, or NULL (default)}

\item{message}{An additional message handler to call before the one
we use internally, or NULL (default)}

\item{warning}{An additional warning handler to call before the one
we use internally, or NULL (default)}

\item{error}{An additional error handler to call before the one we
use internally, or NULL (default)}

\item{value}{An additional value handler to call before the one we
use internally, or NULL (default)}
}
\value{
A function which can be used as an evaluator with the
    RCacheSuite machinery
}
\description{
Create an evaluation function appropiate for use with
    RCacheSuite that uses Wickham's evaluate function as the
    underlying mechanism
}
