% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{mustForce}
\alias{mustForce}
\title{mustForce}
\usage{
mustForce(codeInf, nonCachedFuns = nocache)
}
\arguments{
\item{codeInf}{A ScriptNodeInfo object for the code being
assessed}

\item{nonCachedFuns}{a vector of function names which should be
considered uncacheable}
}
\value{
\code{TRUE} if any of the functions in \code{nonCachedFuns} are
    called, \code{FALSE} otherwise.
}
\description{
A simple heuristic function which returns false
    (i.e., uncacheable) if the code calls any of the functions in
    \code{nonCachedFuns}, which defaults to \code{par},
    \code{options}, \code{load}, \code{rm}, and \code{source}.
}
