% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{methods}
\name{WithVisValue-class}
\alias{WithVisValue-class}
\alias{PlotList-class}
\alias{WithVisPlusGraphics-class}
\alias{CachingEngine-class}
\alias{cacheClass}
\alias{CodeCacheSet-class}
\alias{codeCacheSet}
\alias{CachedData-class}
\alias{cachedData}
\title{ObjectClasses}
\description{
these object classes are used by the RCacheSuite package. most should not be created or interacted with manually.
}

