% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{sameOutVar}
\alias{sameOutVar}
\title{sameOutVar}
\usage{
sameOutVar(code)
}
\arguments{
\item{code}{The code or ScriptNodeInfo to be assessed}
}
\value{
TRUE if one (or less) unique variable is output or updated across all expressions in \code{code}, FALSE if two or more unique variables are.
}
\description{
this function provides a simple heuristic for when a
    block of code should be cached as a single entity. It returns
    true when all expressions in the block create or modify the
    same variable (or none at all), e.g.,
\code{
x=data.frame(y=rnorm(100), z=rnorm(100))
x$y[x$y > 3] = NA
x$col3 = sample(1:4, 100, TRUE)
x = x[1:50,]
}
}
